plugins/LLM/requirements.txt
torch==2.3.1
llama-cpp-python==0.2.79
